+++ examples/react-vite/src/tour/flows.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable react-refresh/only-export-components -- Flow definitions with inline content components */
 
 import type { FlowDefinition } from '@flowsterix/core'
 import { createFlow } from '@flowsterix/core'
@@ -11,7 +10,7 @@ import {
   StepText,
   StepTitle,
 } from '@/components/step-content'
-import { useDelayAdvance } from '@/hooks/use-tour'
+import { useDelayAdvance } from '@flowsterix/react'
 
 import { getTourRouter } from './routerBridge'